requirements.txt
pandas
numpy
tqdm
openai
tenacity
python-dotenv
convokit
